+++ Makefile
@@ -1,7 +1,7 @@
 .PHONY: build clean deploy
 
 build:
-	env GOARCH=amd64 GOOS=linux go build -ldflags="-s -w" -o bin/bot bot/main.go
+	env GOARCH=amd64 GOOS=linux go build -ldflags="-s -w" -o bin/bot bot/*.go
 
 clean:
 	rm -rf ./bin
