+++ bot/main.go
@@ -1,37 +1,71 @@
 package main
 
 import (
-	"errors"
-	"fmt"
+	"crypto/hmac"
+	"crypto/sha256"
+	"encoding/base64"
+	"encoding/json"
+	"log"
 	"os"
 
 	"github.com/aws/aws-lambda-go/events"
 	"github.com/aws/aws-lambda-go/lambda"
-	"github.com/line/line-bot-sdk-go/v7/linebot"
+	"github.com/line/line-bot-sdk-go/linebot"
 )
 
 type Response events.APIGatewayProxyResponse
 
 func Handler(request events.APIGatewayProxyRequest) (events.APIGatewayProxyResponse, error) {
-	line := Line{}
-	err := line.New(
-		os.Getenv("LINE_BOT_CHANNEL_SECRET"),
-		os.Getenv("LINE_BOT_CHANNEL_TOKEN"))
+
+	bot, err := linebot.New(
+		os.Getenv("LINE_CHANNE_TOKEN"),
+		os.Getenv("LINE_CHANNEL_SECRET"),
+	)
+	// エラーに値があればログに出力し終了する
 	if err != nil {
-		fmt.Println(err)
+		log.Fatal(err)
 	}
-	eve, err := ParseRequest(line.ChannelSecret, request)
+	result := "hoge"
+
+	// テキストメッセージを生成する
+	message := linebot.NewTextMessage(result)
+	// テキストメッセージを友達登録しているユーザー全員に配信する
+	if _, err := bot.BroadcastMessage(message).Do(); err != nil {
+		log.Fatal(err)
+	}
+
+	return events.APIGatewayProxyResponse{
+		Body:       "",
+		StatusCode: 200,
+	}, nil
+
+}
+func ParseRequest(channelSecret string, r events.APIGatewayProxyRequest) ([]*linebot.Event, error) {
+	if !ValidateSignature(channelSecret, r.Headers["x-line-signature"], []byte(r.Body)) {
+		return nil, linebot.ErrInvalidSignature
+	}
+	request := &struct {
+		Events []*linebot.Event `json:"events"`
+	}{}
+	if err := json.Unmarshal([]byte(r.Body), request); err != nil {
+		return nil, err
+	}
+	return request.Events, nil
+}
+
+func ValidateSignature(channelSecret string, signature string, body []byte) bool {
+	decoded, err := base64.StdEncoding.DecodeString(signature)
 	if err != nil {
-		status := 200
-		if err == linebot.ErrInvalidSignature {
-			status = 400
-		} else {
-			status = 500
-		}
-		return events.APIGatewayProxyResponse{StatusCode: status}, errors.New("Bat Request")
+		return false
 	}
-	line.EventRouter(eve)
-	return events.APIGatewayProxyResponse{Body: request.Body, StatusCode: 200}, nil
+
+	hash := hmac.New(sha256.New, []byte(channelSecret))
+	_, err = hash.Write(body)
+	if err != nil {
+		return false
+	}
+
+	return hmac.Equal(decoded, hash.Sum(nil))
 }
 
 func main() {
