+++ go.mod
@@ -2,9 +2,9 @@ module github.com/nkchan/nondenbot-reply/bot
 
 go 1.17
 
-require github.com/aws/aws-lambda-go v1.28.0
-
 require (
-	github.com/line/line-bot-sdk-go v7.8.0+incompatible // indirect
-	github.com/line/line-bot-sdk-go/v7 v7.14.0 // indirect
+	github.com/aws/aws-lambda-go v1.29.0
+	github.com/line/line-bot-sdk-go/v7 v7.14.0
 )
+
+require github.com/line/line-bot-sdk-go v7.8.0+incompatible // indirect
